Extract concern for listening to model changes
So that we use it in non-Broadcast classes.

Removed 'suppression' parts as we don't currently use it.
