+++ app/broadcasts/application_broadcast.rb
@@ -1,45 +1,5 @@
 class ApplicationBroadcast
-  include ActiveSupport::Callbacks
-
-  thread_cattr_accessor :suppressed, instance_writer: false
-
-  attr_reader :record, :event
-  alias_method :model, :record
-
-  delegate_missing_to :record
-
-  define_callbacks :update
-
-  def self.suppressed?
-    current = self
-
-    loop do
-      break true if current.suppressed
-      break false if current == ApplicationBroadcast
-
-      current = current.superclass
-    end
-  end
-
-  def self.suppress
-    self.suppressed = true
-    yield
-  ensure
-    self.suppressed = nil
-  end
-
-  def self.changed(*attrs, **options, &block)
-    attrs = attrs.to_set
-
-    options[:if] = [
-      proc { attrs.intersect?(changed_attributes) },
-      *options[:if]
-    ]
-
-    set_callback(:update, *[*options.delete(:call), options], &block)
-  end
-
-  private_class_method :changed
+  include ModelListener
 
   def self.inherited(klass)
     super
@@ -48,40 +8,7 @@ class ApplicationBroadcast
     klass.alias_method model_key, :record
   end
 
-  # Wisper interface
-  %i[created updated destroyed].each do |event|
-    class_eval <<~RUBY, __FILE__, __LINE__ + 1
-      def self.model_#{event}(record, *args)
-        return if suppressed?
-
-        new(record, *args).#{event}
-      rescue => error
-        Rails.logger.error("\#{self} (#{event}) error: \#{error.inspect}")
-        Rails.logger.error(error.backtrace.join("\\n"))
-        Bugsnag.notify(error)
-      end
-    RUBY
-  end
-
-  def initialize(record, changes = nil)
-    @record = record
-    @recorded_changes = changes
-  end
-
-  def created; end
-  def destroyed; end
-
-  def updated
-    run_callbacks(:update)
-  end
-
   private
-
-  attr_reader :recorded_changes
-
-  def changed_attributes
-    @changed_attributes ||= recorded_changes.keys.map(&:to_sym).to_set
-  end
 
   def blueprint(model, *args)
     "#{model.class}Blueprint".constantize.render_as_hash(model, *args)
